fix(events): wrap update_event field edits + status transition in a transaction (Issue 721)
_apply_field_updates() and _handle_status_update() each autocommit via
event.save(), so a failed status transition after a successful field
edit left the field edit persisted despite the client getting a 4xx.
Wraps both in transaction.atomic() so a failed transition rolls back
the field edits too.
